pyproject.toml
[tool.poetry]
name = "algo"
version = "0.1.0"
description = ""
authors = ["kumarvinayak490 <vinayak@ineuron.ai>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.2"
djangorestframework = "^3.14.0"
django-filter = "^23.5"
django-split-settings = "^1.2.0"
pyyaml = "^6.0.1"
djangorestframework-simplejwt = "^5.3.1"
django-cors-headers = "^4.3.1"
redis = "^5.0.1"
celery = "^5.3.6"


[tool.poetry.group.dev.dependencies]
flake8 = "^7.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
